dummy-producer-consumer/requirements.txt
paho-mqtt